Cambios

+++ database/migrations/2023_03_14_220812_create_reservacions_table.php
@@ -26,7 +26,7 @@ return new class extends Migration {
             $table->integer('precioAdultos');
             $table->integer('cantidadNinios')->default(0);
             $table->integer('precioNinios')->default(0);
-            $table->float('anticipo');
+            $table->integer('anticipo');
             $table->boolean('formaPago')->default(1);
             $table->boolean('estado')->default(0);
             $table->timestamps();
